3d/requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
anytree==2.8.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
einops==0.6.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
fonttools==4.37.1
frozenlist==1.3.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.49.0
hopcroftkarp==1.2.5
humanize==4.5.0
idna==3.4
ignite==1.1.0
imageio==2.21.3
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
loguru==0.6.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
MedPy==0.4.0
mistune==2.0.4
monai==0.9.1
multidict==6.0.2
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.6
nibabel==5.0.0
notebook==6.4.12
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
panel==0.13.1
param==1.12.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyct==0.4.8
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-ignite==0.4.10
pytz==2022.7.1
pyvista==0.36.1
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
Rtree==1.0.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
scooby==0.6.0
Send2Trash==1.8.0
shellingham==1.5.0.post1
SimpleITK==2.2.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
svgwrite==1.4.3
tensorboard==2.10.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typer==0.7.0
typing_extensions==4.3.0
urllib3==1.26.12
vtk==9.2.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wslink==1.8.2
yarl==1.8.1